Moved all board render functions to its controller
